+++ src/components/DocPages/TravelPolicy.jsx
@@ -1,0 +1,107 @@
+import React from "react"
+
+import Container from "./Container"
+import Text from "./Text"
+import Content from "../About/Content"
+
+const TravelPolicy = () => (
+  <Container>
+    <Content>
+      <Text>
+        <h1>Travel Policy</h1>
+        <h2>Introduction</h2>
+        <p>
+          At McHacks, we look forward to providing travel accomodation for as
+          many attendees as we can. We care about bringing a great travel
+          experience and so, we have outlined a few rules in this policy to
+          adhere to that mission.
+          <br></br>
+          <br></br>
+          Our travel policy is complementary to the{" "}
+          <a href="/code-of-conduct" target="_top">
+            McHacks Code of Conduct
+          </a>
+          . If you’re an attendee with any travel accommodation, this policy
+          applies alongside our Code of Conduct.
+        </p>
+        <h2>Buses</h2>
+        <h3>Seat Capacity</h3>
+        <p>
+          Bus seats are available at a first-come first-serve basis to all
+          attendees with an acceptance status on our hacker dashboard.
+          <br></br>
+          <br></br>
+          If an attendee can no longer attend, please email{" "}
+          <a href="mailto:contact@mchacks.ca" target="_top">
+            contact@mchacks.ca
+          </a>{" "}
+          to discuss arrangements for a refund so we can open the seat up to
+          other interested attendees.
+        </p>
+        <h3>Deposits</h3>
+        <p>
+          To secure a seat on one of our buses, an attendee must place a
+          monetary deposit at the time of seat security. Bus seats cannot be
+          traded or sold to another attendee or person at anytime.
+          <br></br>
+          <br></br>
+          Bus deposits are refunded to the attendee if they (a) confirm their
+          attendance on our hacker dashboard, (b) check-in to McHacks as a
+          hacker, and (c) submit a project to at least one of our prize
+          categories.
+          <br></br>
+          <br></br>
+          Deposits will be refunded to attendees upon review of the above
+          qualifications and distributed within 15 days after McHacks.
+        </p>
+        <h2>Self-Arranged Travel</h2>
+        <p>
+          Any attendee is welcome to arrange their own travel accommodations to
+          and from McHacks. Liability, safety, and insurance related to
+          self-arranged travel is unaffiliated with McHacks and left to an
+          attendee's discretion.
+        </p>
+        <h3>Travel Offer</h3>
+        <p>
+          On a case-by-case basis, an attendee may be offered a monetary travel
+          offer on their hacker dashboard to support their own travel
+          arrangements to and from McHacks.
+          <br></br>
+          <br></br>
+          The travel offer given to any attendee may be retracted at the McHacks
+          team’s discretion.
+        </p>
+        <h3>Travel Reimbursement</h3>
+        <p>
+          An attendee is eligible for travel reimbursement of the offer amount
+          indicated on their hacker dashboard if they (a) confirm their
+          attendance on our hacker dashboard, (b) check-in to McHacks as a
+          hacker, and (c) submit a project to at least one of our prize
+          categories.
+          <br></br>
+          <br></br>
+          An attendee must upload and provide a digital copy of their travel
+          receipts to be eligible for travel reimbursement. These receipts must
+          be provided to McHacks by February 1, 2020 at 12:00pm EST.
+          <br></br>
+          <br></br>
+          All travel reimbursements must be picked up during McHacks and before
+          February 2, 2020 at 5:00pm EST or an attendee will not be offered any
+          reimbursement.
+        </p>
+        <h2>Questions</h2>
+        <p>
+          Any questions regarding the McHacks Travel Policy may be directed to{" "}
+          <a href="mailto:contact@mchacks.ca" target="_top">
+            contact@mchacks.ca
+          </a>
+          .<br></br>
+          <br></br>
+          <i>Last modified on January 13, 2020.</i>
+        </p>
+      </Text>
+    </Content>
+  </Container>
+)
+
+export default TravelPolicy