+++ docs/CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.4.0](https://github.com/hackmcgill/mchacks7/tree/1.4.0) - 2020-01-14
+
+### Added
+
+- Add travel policy page
+
+### Fixed
+
+- Add modified date to privacy policy
+
 ## [1.3.2](https://github.com/hackmcgill/mchacks7/tree/1.3.2) - 2019-12-22
 
 ### Added